device: prevent more races on endpoint
The existing code attempted to prevent races on peer.endpoint
due to logging by adding a no-op String method to conn.NativeEndpoint.
However, the conn.NativeEndpoint got shoved inside a conn.Endpoint
interface, which meant that there was already a race trying to get
the type out in order to discover that printing it was a no-op.

Improve the bandaid by reading peer.endpoint entirely
in the critical section.

Even better yet would be to make conn.NativeEndpoint a struct
that does some internal synchronization, but that would be
a much more invasive change.