conn: set SO_{SND,RCV}BUF to 7MB on the Bind UDP socket
The conn.Bind UDP socket send and receive buffers are now being sized to
7MB, whereas they were previously inheriting the system defaults. The
value of 7MB is chosen as it is the max supported by a default
configuration of macOS. Some platforms will silently clamp the value to
other maximums. On Linux, we use SO_{SND,RCV}BUFFORCE in case 7MB is
beyond net.core.{r,w}mem_max.

Signed-off-by: Jordan Whited <jordan@tailscale.com>
Signed-off-by: James Tucker <james@tailscale.com>
Co-authored-by: James Tucker <james@tailscale.com>